修复输入部分字符导致的崩溃问题
